cleanup exported api
- timestamps are now always int32
- export fields from response structs

+++ demo/main.go
@@ -39,7 +39,7 @@ func main() {
 
 	scanner := bufio.NewScanner(f)
 
-	w := carbonmem.NewWhisper(int64(*epoch0), *wsize)
+	w := carbonmem.NewWhisper(int32(*epoch0), *wsize)
 
 	var lines int
 
@@ -58,7 +58,7 @@ func main() {
 			log.Println("processed", lines)
 		}
 
-		w.Incr(int64(t), fields[1], 1)
+		w.Incr(int32(t), fields[1], 1)
 	}
 
 	var m runtime.MemStats
